+++ bin-proto/src/discriminable.rs
@@ -1,0 +1,5 @@
+pub trait Discriminable {
+    type Discriminant;
+
+    fn discriminant(&self) -> Self::Discriminant;
+}